% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmodel.R
\docType{package}
\name{stmodel-package}
\alias{stmodel}
\alias{stmodel-package}
\title{stmodel: Fit Spatio-Temporal Linear Mixed Models}
\description{
A R package for fitting spatio-temporal models. The functions most often
used are \code{stlmm()} for fitting spatio-temporal linear mixed
models, \code{summary()} for summarizing those models, and \code{predict}
for making predictions.
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Authors:
\itemize{
  \item Matt Higham \email{mhigham@stlawu.edu}
}

}
